refactor: rename TalkControlMaster to TCController
