+++ src/client/layouts/presenter/index.js
@@ -1,13 +1,13 @@
 'use strict';
 
-import { TalkControlMaster } from '../../talk-control-master/talk-control-master.js';
+import { TCController } from '@client/tc-controller/tc-controller.js';
 import config from '@config/config.json';
 
 window.addEventListener('DOMContentLoaded', function() {
     const isRemote = window.location.href.indexOf('://localhost:') === -1;
 
-    const talkControlMaster = new TalkControlMaster(isRemote ? config.tcServer.urls.external : config.tcServer.urls.local);
-    talkControlMaster.init();
+    const tcController = new TCController(isRemote ? config.tcServer.urls.external : config.tcServer.urls.local);
+    tcController.init();
 
     const url = sessionStorage.getItem('presentationUrl');
     if (url) {
